test(logic): extract C5 riches test helpers to keep files <=400 lines
Splits shared fixtures from the two purchased-tile riches test files into
sibling *_test_support.dart files so each test file stays at or below the
repo.logic_test_file_size 400-physical-line limit enforced by ct_repo_lint.

No behaviour change. All 24 tests still pass via the public helper names.

Refs #2991
