+++ packages/colonizethis_logic/test/turn/riches_to_treasury_phase_purchased_tile_riches_test.dart
@@ -15,6 +15,7 @@ import 'package:colonizethis_models/colonizethis_models.dart';
 import 'package:colonizethis_test/test.dart';
 
 import '../test_fixtures.dart';
+import 'riches_to_treasury_phase_purchased_tile_riches_test_support.dart';
 
 void main() {
   group('richesToTreasuryTurnPhaseHandler — purchased-tile riches handoff', () {
@@ -23,7 +24,7 @@ void main() {
       'minor province credits owning GP treasury without affecting the '
       "GP's own stockpile riches conversion",
       () {
-        final game = _gameWithPurchasedGoldTile(
+        final game = gameWithPurchasedGoldTile(
           gpATreasury: 100,
           gpAStockpileGold: 0,
         );
@@ -31,7 +32,7 @@ void main() {
         final config = TurnResolverConfig(
           topology: const MapTopology(nodes: [], edges: []),
           orders: const Orders(),
-          tileMapByRegion: _tileMapByRegion(Resource.gold),
+          tileMapByRegion: tileMapByRegionForResource(Resource.gold),
         );
 
         final next =
@@ -58,7 +59,7 @@ void main() {
       'GP own-stockpile riches still convert at the same time as the '
       'purchased-tile credit (no regression)',
       () {
-        final game = _gameWithPurchasedGoldTile(
+        final game = gameWithPurchasedGoldTile(
           gpATreasury: 100,
           gpAStockpileGold: 2,
         );
@@ -66,7 +67,7 @@ void main() {
         final config = TurnResolverConfig(
           topology: const MapTopology(nodes: [], edges: []),
           orders: const Orders(),
-          tileMapByRegion: _tileMapByRegion(Resource.gold),
+          tileMapByRegion: tileMapByRegionForResource(Resource.gold),
         );
 
         final next =
@@ -87,7 +88,7 @@ void main() {
       'AC purchased-tile riches handoff — non-riches resource: timber tile '
       'produces no purchased-tile credit (stockpile riches still convert)',
       () {
-        final game = _gameWithPurchasedTileResource(
+        final game = gameWithPurchasedTileResource(
           resource: Resource.timber,
           improvementLevel: 1,
           roadLevel: 1,
@@ -98,7 +99,7 @@ void main() {
         final config = TurnResolverConfig(
           topology: const MapTopology(nodes: [], edges: []),
           orders: const Orders(),
-          tileMapByRegion: _tileMapByRegion(Resource.timber),
+          tileMapByRegion: tileMapByRegionForResource(Resource.timber),
         );
 
         final next =
@@ -118,7 +119,7 @@ void main() {
       'AC purchased-tile riches handoff — unimproved tile: improvement '
       'level 0 produces no credit',
       () {
-        final game = _gameWithPurchasedTileResource(
+        final game = gameWithPurchasedTileResource(
           resource: Resource.silver,
           improvementLevel: 0,
           roadLevel: 1,
@@ -129,7 +130,7 @@ void main() {
         final config = TurnResolverConfig(
           topology: const MapTopology(nodes: [], edges: []),
           orders: const Orders(),
-          tileMapByRegion: _tileMapByRegion(Resource.silver),
+          tileMapByRegion: tileMapByRegionForResource(Resource.silver),
         );
 
         final next =
@@ -186,7 +187,7 @@ void main() {
         final config = TurnResolverConfig(
           topology: const MapTopology(nodes: [], edges: []),
           orders: const Orders(),
-          tileMapByRegion: _tileMapByRegion(Resource.gold),
+          tileMapByRegion: tileMapByRegionForResource(Resource.gold),
         );
 
         final next =
@@ -210,7 +211,7 @@ void main() {
       'minor seller is never credited — `Game.players` does not gain a '
       'minor entry from the riches handoff',
       () {
-        final game = _gameWithPurchasedGoldTile(
+        final game = gameWithPurchasedGoldTile(
           gpATreasury: 0,
           gpAStockpileGold: 0,
         );
@@ -218,7 +219,7 @@ void main() {
         final config = TurnResolverConfig(
           topology: const MapTopology(nodes: [], edges: []),
           orders: const Orders(),
-          tileMapByRegion: _tileMapByRegion(Resource.gold),
+          tileMapByRegion: tileMapByRegionForResource(Resource.gold),
         );
 
         final next =
@@ -241,7 +242,7 @@ void main() {
       'no `tileMapByRegion` on config — handler is a no-op for purchased '
       'tile credits and existing GP stockpile riches still convert',
       () {
-        final game = _gameWithPurchasedGoldTile(
+        final game = gameWithPurchasedGoldTile(
           gpATreasury: 100,
           gpAStockpileGold: 2,
         );
@@ -275,7 +276,7 @@ void main() {
       'matches the behaviour of regular riches-to-treasury cash-in',
       () {
         // El Dorado scenario uses 1.5×.
-        final game = _gameWithPurchasedTileResource(
+        final game = gameWithPurchasedTileResource(
           resource: Resource.spices,
           improvementLevel: 1,
           roadLevel: 1,
@@ -287,7 +288,7 @@ void main() {
         final config = TurnResolverConfig(
           topology: const MapTopology(nodes: [], edges: []),
           orders: const Orders(),
-          tileMapByRegion: _tileMapByRegion(Resource.spices),
+          tileMapByRegion: tileMapByRegionForResource(Resource.spices),
         );
 
         final next =
@@ -335,7 +336,7 @@ void main() {
 
   group('applyPurchasedTileRichesHandoff — direct helper', () {
     test('null tileMapByRegion → identity (no-op)', () {
-      final game = _gameWithPurchasedGoldTile(gpATreasury: 99);
+      final game = gameWithPurchasedGoldTile(gpATreasury: 99);
       final result = applyPurchasedTileRichesHandoff(
         game,
         tileMapByRegion: null,
@@ -344,7 +345,7 @@ void main() {
     });
 
     test('empty tileMapByRegion → identity (no-op)', () {
-      final game = _gameWithPurchasedGoldTile(gpATreasury: 99);
+      final game = gameWithPurchasedGoldTile(gpATreasury: 99);
       final result = applyPurchasedTileRichesHandoff(
         game,
         tileMapByRegion: const <String, TileMapResult>{},
@@ -357,83 +358,9 @@ void main() {
       final game = TestFixtures.minimalGame();
       final result = applyPurchasedTileRichesHandoff(
         game,
-        tileMapByRegion: _tileMapByRegion(Resource.gold),
+        tileMapByRegion: tileMapByRegionForResource(Resource.gold),
       );
       expect(identical(result, game), isTrue);
     });
   });
-}
-
-Game _gameWithPurchasedGoldTile({
-  required int gpATreasury,
-  int gpAStockpileGold = 0,
-}) =>
-    _gameWithPurchasedTileResource(
-      resource: Resource.gold,
-      improvementLevel: 1,
-      roadLevel: 1,
-      gpATreasury: gpATreasury,
-      gpAStockpileGold: gpAStockpileGold,
-    );
-
-Game _gameWithPurchasedTileResource({
-  required Resource resource,
-  required int improvementLevel,
-  required int roadLevel,
-  required int gpATreasury,
-  required int gpAStockpileGold,
-  double richesCashMultiplier = 1.0,
-}) {
-  const ow = 'oldWorld';
-  const minorProvinceId = '$ow|M1';
-  const tileKey = '$ow|M1|0|0';
-  TileMapState tileState = const TileMapState();
-  if (improvementLevel > 0) {
-    tileState = tileState.setImprovement(tileKey, improvementLevel);
-  }
-  if (roadLevel > 0) {
-    tileState = tileState.setRoadLevel(tileKey, roadLevel);
-  }
-  return TestFixtures.minimalGame(
-    players: [
-      Player(
-        id: 'gpA',
-        displayName: 'GP A',
-        isHuman: true,
-        treasury: gpATreasury,
-        stockpile: gpAStockpileGold > 0
-            ? Stockpile(quantities: {'gold': gpAStockpileGold})
-            : Stockpile.empty,
-      ),
-    ],
-    oldWorld: const RegionData(
-      provinces: [
-        Province(id: minorProvinceId, regionId: ow, ownerId: 'M1'),
-      ],
-    ),
-    tileKeysByRegionAndProvince: const {
-      ow: {
-        minorProvinceId: [tileKey],
-      },
-    },
-    minorNations: const [MinorNation(id: 'M1', displayName: 'Minor 1')],
-    purchasedTilesByTileKey: const {tileKey: 'gpA'},
-    tileState: tileState,
-    richesCashMultiplier: richesCashMultiplier,
-  );
-}
-
-Map<String, TileMapResult> _tileMapByRegion(Resource resource) {
-  return {
-    'oldWorld': TileMapResult(
-      width: 1,
-      height: 1,
-      grid: [
-        ['M1'],
-      ],
-      resourceGrid: [
-        [resource],
-      ],
-    ),
-  };
 }
